+++ spec/main-spec.js
@@ -1,143 +1,145 @@
 'use babel'
 
-import {RangePool} from '../'
-console = require('console')
+import RangePool from '../'
 
 describe('RangePool', function() {
+  function getRangePool(limit: number): RangePool {
+    return new RangePool(limit)
+  }
+
   it('cries when constructor param is inifinite', function() {
     expect(function() {
-      new RangePool(Infinity)
+      getRangePool(Infinity)
     }).toThrow()
   })
   it('cries when constructor param is negative', function() {
     expect(function() {
-      new RangePool(-1)
+      getRangePool(-1)
     }).toThrow()
   })
   it('accepts a number', function() {
-    const pool = new RangePool(50)
+    const pool = getRangePool(50)
     expect(pool).toBeDefined()
   })
 
-  it('can divide a workaround in a lot of workers', function() {
-    const pool = new RangePool(500)
-    let i =0
+  it('can divide a workload in a lot of workers', function() {
+    const pool = getRangePool(500)
+    let i = 0
     for (i = 0; i < 10; i++) {
-      pool.createWorker()
+      pool.getWorker()
     }
   })
   it('cries if we try to create a worker on a completed pool', function() {
-    const pool = new RangePool(500)
-    const worker = pool.createWorker()
-    worker.advance(500)
+    const pool = getRangePool(500)
+    pool.getWorker().advance(500)
     expect(function() {
-      pool.createWorker()
+      pool.getWorker()
     }).toThrow()
   })
   it('has a working hasCompleted method', function() {
-    const pool = new RangePool(500)
-    const worker = pool.createWorker()
+    const pool = getRangePool(500)
+    const worker = pool.getWorker()
     expect(pool.hasCompleted()).toBe(false)
     worker.advance(500)
     expect(pool.hasCompleted()).toBe(true)
   })
-  it('has a working getCompletedSteps method', function() {
-    const pool = new RangePool(500)
-    expect(pool.getCompletedSteps()).toBe(0)
-    const worker = pool.createWorker()
-    expect(pool.getCompletedSteps()).toBe(0)
+  it('has a working getCompleted method', function() {
+    const pool = getRangePool(500)
+    expect(pool.getCompleted()).toBe(0)
+    const worker = pool.getWorker()
+    expect(pool.getCompleted()).toBe(0)
     worker.advance(50)
-    expect(pool.getCompletedSteps()).toBe(50)
+    expect(pool.getCompleted()).toBe(50)
   })
   it('has a working getRemaining method', function() {
-    const pool = new RangePool(500)
+    const pool = getRangePool(500)
     expect(pool.getRemaining()).toBe(500)
-    const worker = pool.createWorker()
+    const worker = pool.getWorker()
     expect(pool.getRemaining()).toBe(500)
     worker.advance(100)
     expect(pool.getRemaining()).toBe(400)
   })
 
   it('properly distributes work among workers', function() {
-    const pool = new RangePool(512)
-    const workerFirst = pool.createWorker()
-    expect(pool.getCompletedSteps()).toBe(0)
+    const pool = getRangePool(512)
+    const workerFirst = pool.getWorker()
+    expect(pool.getCompleted()).toBe(0)
     workerFirst.advance(256)
-    expect(pool.getCompletedSteps()).toBe(256)
-    const workerSecond = pool.createWorker()
+    expect(pool.getCompleted()).toBe(256)
+    const workerSecond = pool.getWorker()
     expect(workerFirst.limitIndex).toBe(384)
     expect(workerFirst.currentIndex).toBe(256)
     expect(workerSecond.startIndex).toBe(384)
     expect(workerSecond.currentIndex).toBe(384)
     expect(workerSecond.limitIndex).toBe(512)
-    expect(pool.getCompletedSteps()).toBe(256)
+    expect(pool.getCompleted()).toBe(256)
     workerSecond.advance(128)
-    expect(pool.getCompletedSteps()).toBe(384)
+    expect(pool.getCompleted()).toBe(384)
   })
 
   it('properly distributes work among more than two workers', function() {
-    const pool = new RangePool(1024)
-    const workerFirst = pool.createWorker()
+    const pool = getRangePool(1024)
+    const workerFirst = pool.getWorker()
     workerFirst.advance(128)
-    const workerSecond = pool.createWorker()
+    const workerSecond = pool.getWorker()
     workerSecond.advance(100)
-    const workerThird = pool.createWorker()
-    expect(pool.getCompletedSteps()).toBe(228)
+    const workerThird = pool.getWorker()
+    expect(pool.getCompleted()).toBe(228)
 
     let range = 0
-    range += workerFirst.getIndexLimit() - workerFirst.getStartIndex()
-    range += workerSecond.getIndexLimit() - workerSecond.getStartIndex()
-    range += workerThird.getIndexLimit() - workerThird.getStartIndex()
+    range += workerFirst.getLimitIndex() - workerFirst.getStartIndex()
+    range += workerSecond.getLimitIndex() - workerSecond.getStartIndex()
+    range += workerThird.getLimitIndex() - workerThird.getStartIndex()
     expect(range).toBe(1024)
   })
 
   it('properly distributes work of odd length', function() {
-    const pool = new RangePool(999)
-    const workerFirst = pool.createWorker()
+    const pool = getRangePool(999)
+    const workerFirst = pool.getWorker()
     workerFirst.advance(50)
-    const workerSecond = pool.createWorker()
+    const workerSecond = pool.getWorker()
     expect(workerFirst.limitIndex).toBe(525)
     expect(workerSecond.startIndex).toBe(525)
     expect(workerSecond.limitIndex).toBe(999)
   })
 
   it('re-uses old unfinished died workers even if that means one', function() {
-    const pool = new RangePool(90)
-    const workerA = pool.createWorker()
+    const pool = getRangePool(90)
+    const workerA = pool.getWorker()
     workerA.advance(50)
     workerA.dispose()
-    expect(workerA.isActive()).toBe(false)
-    const workerB = pool.createWorker()
+    expect(workerA.getActive()).toBe(false)
+    const workerB = pool.getWorker()
     expect(workerA).toBe(workerB)
-    expect(workerB.isActive()).toBe(true)
+    expect(workerB.getActive()).toBe(true)
   })
 
   it('re-uses old unfinished died workers no matter how many', function() {
-    const pool = new RangePool(50)
-    const workerA = pool.createWorker()
+    const pool = getRangePool(50)
+    const workerA = pool.getWorker()
     workerA.advance(5)
-    const workerB = pool.createWorker()
+    const workerB = pool.getWorker()
     workerB.advance(5)
-    const workerC = pool.createWorker()
-    expect(workerC.isActive()).toBe(true)
+    const workerC = pool.getWorker()
+    expect(workerC.getActive()).toBe(true)
     workerC.dispose()
-    expect(workerC.isActive()).toBe(false)
-    const workerD = pool.createWorker()
+    expect(workerC.getActive()).toBe(false)
+    const workerD = pool.getWorker()
     expect(workerD).toBe(workerC)
-    expect(workerD.isActive()).toBe(true)
-    expect(workerC.isActive()).toBe(true)
+    expect(workerD.getActive()).toBe(true)
+    expect(workerC.getActive()).toBe(true)
   })
 
   it('is serializable', function() {
-    const pool = new RangePool(50)
-    const workerA = pool.createWorker()
+    const pool = getRangePool(50)
+    const workerA = pool.getWorker()
     workerA.advance(5)
-    const workerB = pool.createWorker()
+    const workerB = pool.getWorker()
     workerB.advance(5)
 
     const poolClone = RangePool.unserialize(pool.serialize())
     expect(pool.length).toEqual(poolClone.length)
-    expect(pool.complete).toEqual(poolClone.complete)
+    expect(pool.hasCompleted()).toEqual(poolClone.hasCompleted())
     expect([...pool.workers]).toEqual([...poolClone.workers])
   })
 })