+++ src/types.js
@@ -1,0 +1,8 @@
+/* @flow */
+
+export type SerializedWorker = {
+  active: boolean,
+  startIndex: number,
+  limitIndex: number,
+  currentIndex: number
+}